Parses commandline help options to return vector of valid flag names

When using cmdfun to write lazy shell wrappers, the user can easily mistype a commandline flag since there is not text completion. Some programs behave unexpectedly when flags are typed incorrectly, and for this reason return uninformative error messages.

cmd_help_parse_flags(help_lines, processx = FALSE)

Arguments

help_lines
character vector contianing the output of "command --help", or similar output. Optional: pass either stdout, or stderr output from processx::run(), must set processx = TRUE.

processx
logical(1) if set to TRUE will split string on "\n" before parsing.

Value

first word entry of each line prefixed by "-". typically this grabs all flags from --help output, but often times will require extra postprocessing to remove commas or equal signs, for example.

Details

cmd_help_parse_flags tries to grab flags from --help documentation which can be used for error checking.

See also

Examples

if (FALSE) {
# with processx
out <- processx::run("tar", "--help", error_on_status = F)
lines <- strsplit(out$stderr, "\n")[[1]]
fn_flags <- cmd_help_parse_flags(lines)

# with system2
lines <- system2("tar", "--help", stderr = T)
fn_flags <- cmd_help_parse_flags(lines)
}